src/requirements.txt
flask==2.0.1
werkzeug==2.0.2
redis==4.3.4
PyJWT==2.1.0
flasgger==0.9.5
prometheus-flask-exporter==0.18.2
flask-redis==0.4.0
gunicorn==20.1.0